requirements.txt
requests
fastmcp>=0.3.0
mcp
python-dotenv
httpx